Avoid arithmetic left shift overflow in reallocarray
Some platforms (WIN64) have a 32-bit long, so shifting 1UL will
cause an overflow. This replaces the 1UL with (size_t)1 so we get
the correct size.

+++ crypto/compat/reallocarray.c
@@ -37,7 +37,7 @@ void *reallocarray(void *ptr, size_t newmem, size_t size)
 
 #else
 
-#define SQRT_SIZE_MAX ((1UL << (sizeof(size_t) * 8 / 2)) - 1)
+#define SQRT_SIZE_MAX (((size_t)1 << (sizeof(size_t) * 8 / 2)) - 1)
 
 void *reallocarray(void *ptr, size_t newmem, size_t size)
 {
